+++ docs/code/types.md
@@ -6,16 +6,16 @@
 [E]: ../icons/empty.png
 [N]: ../icons/na.png
 
-version 1.1.8
+version 1.1.9 [![P]][M]
 ---
 
-| **ID** | элемент        | tools           | версия | mingw 830, 720 710 | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008             |  
-|:------:|:--------------:|:---------------:|:------:|:------------------:|:---------------------:|:---------------------------------------:|  
-|  0000  | [fixed][01]    | 0.1.8           | 1.0.1  |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][VS-OLD]                          |  
-|  0001  | [traits][02]   | 0.1.0           | 0.1.1  |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][0] [![D]][0] [![D]][0] [![D]][0] |  
-|  0002  | [common][03]   | 0.2.6           | 1.0.6  |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![D]][0] [![N]][0] [![N]][0] [![N]][0] |  
-|  0003  | [variadic][04] | 0.2.7 [![P]][M] | 0.0.2  |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][0] [![V]][0] [![V]][0] [![D]][1] |  
-|  0004  | [demangle][05] | 0.2.7 [![P]][M] | 0.0.1  |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][VS-OLD]                          |  
+| **ID** | элемент        | tools           | версия          | mingw 830, 720 710 | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008             |  
+|:------:|:--------------:|:---------------:|:---------------:|:------------------:|:---------------------:|:---------------------------------------:|  
+|  0000  | [fixed][01]    | 0.1.8           | 1.0.1           |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][VS-OLD]                          |  
+|  0001  | [traits][02]   | 0.1.0           | 0.1.2 [![P]][M] |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][0] [![D]][0] [![D]][0] [![D]][0] |  
+|  0002  | [common][03]   | 0.2.6           | 1.0.7 [![P]][M] |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![D]][0] [![N]][0] [![N]][0] [![N]][0] |  
+|  0003  | [variadic][04] | 0.2.7 [![P]][M] | 0.0.2           |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][0] [![V]][0] [![V]][0] [![D]][1] |  
+|  0004  | [demangle][05] | 0.2.7 [![P]][M] | 0.0.1           |   [![V]][MINGW]    |  [![V]][VS-NEW]       | [![V]][VS-OLD]                          |  
 
 [M]:  #types  "обработака типов"  
 [MINGW]:   #mingw-new   "поддержка компиляторов mingw"
