+++ docs/code/types/common.md
@@ -6,9 +6,30 @@
 [E]: ../../icons/empty.png
 [N]: ../../icons/na.png
 
-version 1.0.6
+version 1.0.7 [![P]][M]
 ---
 
+common.hpp
+---
+
+| **ID** | элементы           | tools           | commnon | версия | mingw 830, 720 710 | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008             |  
+|:------:|:------------------:|:---------------:|:-------:|:------:|:------------------:|:---------------------:|:---------------------------------------:|  
+|  0200  | [type_of_enum][01] | 0.2.7 [![P]][M] |  1.0.7  | 1.0.0  | [![V]][MINGW]      |  [![V]][VS-NEW]       | [![V]][M] [![N]][4] [![N]][4] [![N]][4] |  
+
+[![D]][M] ограничения
+---------------------
+
+```
+if !dHAS_ENUM_CLASS  (msvc2012)
+    type_of_enum: функциональность не доступна
+
+if !dHAS_USING_ALIAS (msvc2013)
+    type_of_enum_t: функциональность не доступна
+
+```
+
+<br/>
+<br/>
 
 common-msvc2015.hpp
 ---
@@ -100,6 +121,7 @@ small_array
 [1]:       #-is_functor      "ограничение в логике работы для старых компиляторов msvc"  
 [2]:       #-концепты        "ограничение для старых компиляторов msvc"  
 [3]:       #-dereferencable  "ограничение в логике работы для старых компиляторов msvc"  
+[4]:       #-ограничения     "требуется минимум msvc2012"  
 
 
 [00]: #degradate            "метафункция: true, если типы идентичны"
