+++ docs/history.md
@@ -7,9 +7,17 @@
 
 **История изменений**  
 
+
+**Версия 1.1.9 (предварительная)**  
+----------------------------------
+[2021y-04m-03d][23:09:37] #41_stage
+
+1) type_of_enum переехал из traits в common
+
+
 **Версия 1.1.8 (предварительная)**  
 ----------------------------------
-[2021y-04m-03d][23:09:37] #35_traits_4
+[2021y-04m-03d][23:09:37] #39_traits_4
 
 1) обновлен: version                    118
 2) обновлен: types                      118
