+++ docs/table/table.md
@@ -32,24 +32,24 @@ Tools`s version 0.2.7  [![P]][M]
 Разделы
 ---
 
-| **ID** | разделы         |  tools | версия | mingw 830, 720 710  | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008 |  
-|:------:|:---------------:|:------:|:------:|:-------------------:|:---------------------:|:---------------------------:|  
-|  0001  | [features][01]  | 0.0.1  | 1.1.7  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0018  | [macro][18]     | 0.1.7  | 1.0.0  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0019  | [types][19]     | 0.1.8  | 1.1.8  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![D]][VS-OLD]              |  
+| **ID** | разделы         |  tools | версия          | mingw 830, 720 710  | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008 |  
+|:------:|:---------------:|:------:|:---------------:|:-------------------:|:---------------------:|:---------------------------:|  
+|  0001  | [features][01]  | 0.0.1  | 1.1.7           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0018  | [macro][18]     | 0.1.7  | 1.0.0           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0019  | [types][19]     | 0.1.8  | 1.1.9 [![P]][M] |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![D]][VS-OLD]              |  
 
 <br/>
 
 Модули  
 ---
 
-| **ID** | модули          | tools | версия | mingw 830, 720 710  | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008 |  
-|:------:|:---------------:|:-----:|:------:|:-------------------:|:---------------------:|:---------------------------:|  
-|  0000  | [assert][00]    | 0.0.1 | 1.0.1  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0001  | [features][01]  | 0.0.1 | 1.1.7  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0007  | [dFUNCTION][07] | 0.0.6 | 1.0.0  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0018  | [macro][18]     | 0.1.7 | 1.0.0  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
-|  0019  | [types][19]     | 0.1.8 | 1.1.8  |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![D]][VS-OLD]              |  
+| **ID** | модули          | tools | версия          | mingw 830, 720 710  | msvc 2019, 2017, 2015 | msvc 2013, 2012, 2010, 2008 |  
+|:------:|:---------------:|:-----:|:---------------:|:-------------------:|:---------------------:|:---------------------------:|  
+|  0000  | [assert][00]    | 0.0.1 | 1.0.1           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0001  | [features][01]  | 0.0.1 | 1.1.7           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0007  | [dFUNCTION][07] | 0.0.6 | 1.0.0           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0018  | [macro][18]     | 0.1.7 | 1.0.0           |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![V]][VS-OLD]              |  
+|  0019  | [types][19]     | 0.1.8 | 1.1.9 [![P]][M] |    [![V]][MINGW]    |    [![V]][VS-NEW]     | [![D]][VS-OLD]              |  
 
 <br/>
