+++ docs/tasks/2021y-04m-04d-0041-stage.md
@@ -10,7 +10,7 @@ stage
 
 План:  
   - `Часть-1` перестановка:  
-    - [ ] `traits/type_of_enum` переезжает в `types/common`  
+    - [x] `traits/type_of_enum` переезжает в `types/common`  
   - `Часть-1` аудит кода:  
     - [ ] `fixed`  
     - [ ] `common`
