+++ docs/test/200-types.md
@@ -1,6 +1,8 @@
 
 test-200-types
 ---
+  - [x] `common версии 107 [027 PRE]`  
+    - [x] type_of_enum (msvc2012)  
   - [x] `common версии 106 [027 PRE]`  
     - [x] add_const_data  
   - [x] `common версии 105 [027 PRE]`  
@@ -23,6 +25,8 @@ test-200-types
 <br />
 <br />
 
+  - [x] `traits версии 12 [027 PRE]`  
+    - [x] removed: type_of_enum (msvc2012)  
   - [x] `traits версии 11 [027 PRE]`  
      - [x] bugfix: is_class
      - [x] bugfix: is_pointer
